Added message for unsupported image formats

+++ WebAppBuilder/build_new_image.sh
@@ -2,19 +2,25 @@
 
 LOG_OUT=log/stdout.log
 LOG_ERR=log/stderr.log
-exec 1> >(
-	while read -r l; do echo "[$(date +"%Y-%m-%d %H:%M:%S")] $l"; done \
-		| tee -a $LOG_OUT
-	)
-exec 2> >(
-	while read -r l; do echo "[$(date +"%Y-%m-%d %H:%M:%S")] $l"; done \
-	| tee -a $LOG_ERR 1>&2
-	)
+# exec 1> >(
+# 	while read -r l; do echo "[$(date +"%Y-%m-%d %H:%M:%S")] $l"; done \
+# 		| tee -a $LOG_OUT
+# 	)
+# exec 2> >(
+# 	while read -r l; do echo "[$(date +"%Y-%m-%d %H:%M:%S")] $l"; done \
+# 	| tee -a $LOG_ERR 1>&2
+# 	)
 
 baseDir=`dirname $0`
 modelDir=$1
 context=${baseDir}/webAppFiles
 
+if [ ! -e ${modelDir}/model_weights.pth  ] \
+	|| [ ! -e ${modelDir}/model_definition.py  ] \
+	|| [ ! -e ${modelDir}/model_definition.py  ]; then
+	"Files required to build webapp image do not exist in ${modelDir}."
+	exit 404
+fi
 cp --force ${modelDir}/* ${context}/app
 
 Registry=localhost:5050
@@ -27,4 +33,4 @@ echo "docker image build with tag ${Registry}/mywebapp:${timestump}"
 docker image rm -f $(docker images ${Registry}/mywebapp | grep none | awk '{print $3}') 2>/dev/null
 docker images ${Registry}/mywebapp
 
-return $build_result
+exit $build_result
